+++ vue3-apollo-quasar/src/components/FileExplorer/FileExplorerNav.stories.ts
@@ -1,4 +1,5 @@
 import FileExplorerNav from './FileExplorerNav.vue';
+import { ExplorerContent } from './types';
 
 export default {
   title: 'component/FileExplorer/NavItem',
@@ -12,15 +13,7 @@ export default {
   },
 };
 
-type ArgType = {
-  isDirectory: boolean;
-  name: string;
-  latestCommitMessage: string;
-  lastUpdated: string;
-  to: string;
-};
-
-const Template = (args: ArgType) => ({
+const Template = (args: ExplorerContent) => ({
   components: { FileExplorerNav },
   setup() {
     return { args };
