+++ vue3-apollo-quasar/src/components/FileExplorer/types.ts
@@ -1,0 +1,7 @@
+export type ExplorerContent = {
+  isDirectory: boolean;
+  name: string;
+  latestCommitMessage: string;
+  lastUpdated: string;
+  to: string;
+};
